Basic implementation of Poll and Options
Adds the Poll and PollOption content type to provide the ability to
create and manage polls.
A Poll can either be rendered as a page or as a widget.
In this version, only one type of poll is supported: single choice.
Also, only one vote per IP address is allowed. This is hardcoded, but
will be configurable in the future, including voting for authenticated
users.